--- packages/hono-takibi/src/cli/utils/slice-argv.ts
@@ -1,6 +1,0 @@
-/**
- * @param argv - The command line arguments array
- */
-export function sliceArgv(argv: readonly string[]): string[] {
-  return argv.slice(2)
-}
